Add fetching of compiled schema fixtures
This allows the unit tests to fetch compiled schema files in binary
format to perform unit tests on. This is implemented this way so that
we will be able to perform the same unit tests in the browser. We
ideally want all of our parsing and generation code to work in the
browser so we can auto-generate API code if necessary