Read stdin into an ArrayBuffer
This allows the main function to process ArrayBuffers which are
available in Browser engines. The binary script contains Node
specific stuff and everything under the main library function
should work in the browser. It probably makes sense to not have
a main function and just have the helper functions such as
`generate` which creates the AST for the ECMAscript modules
that can be converted with the `toString` API